requirements.txt
python-dotenv==1.0.0
beepy==1.0.7
openai==1.4.0
deepgram-sdk==2.12.0
rhasspy-silence==0.4.0
elevenlabs==0.2.27
pygame==2.5.2

pydub
sounddevice
wave 
timefile
speech_recognition
pyttsx3

# taipy==3.0.0
# pycamera==0.3.2
# opencv-python==4.9.0.80
# ipython==7.18.1

